# src/swagger/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "ApexBridge REST API",
    "version": "1.0.0",
    "description": "High-yield crypto and investment portal REST API specification. Supports complete authentication, live wallet tracking, deposit/withdrawal lifecycles, and yield management.",
    "contact": {
      "name": "ApexBridge Engineering",
      "email": "support@apexbridge.io",
      "url": "https://api.apexbridge.io"
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "Current Server (Local or Cloud)"
    },
    {
      "url": "https://api.apexbridge.io",
      "description": "ApexBridge Production Base URL"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter JWT token obtained from `/api/v1/auth/login` or `/api/v1/auth/signup`."
      }
    },
...
  },
  "paths": {
...
    "/api/v1/user/profile": {
      "get": {
        "summary": "Get user settings, security status, and KYC verification tier",
        "tags": ["2. User Profile & Settings"],
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": {
            "description": "User profile details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": { "type": "string", "example": "usr_8829104" },
                    "name": { "type": "string", "example": "Alexander Vance" },
                    "email": { "type": "string", "example": "alexander@apexbridge.com" },
                    "phone": { "type": "string", "example": "+1 (555) 234-5678" },
                    "tier": { "type": "string", "example": "Tier 2 - Verified" },
                    "is2FAEnabled": { "type": "boolean", "example": true },
                    "currencyPreference": { "type": "string", "example": "USD" },
                    "notifications": {
                      "type": "object",
                      "properties": {
                        "email": { "type": "boolean", "example": true },
                        "sms": { "type": "boolean", "example": false },
                        "yieldAlerts": { "type": "boolean", "example": true }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Update user profile and security preferences",
        "tags": ["2. User Profile & Settings"],
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string", "example": "Alexander Vance" },
                  "phone": { "type": "string", "example": "+1 (555) 234-5678" },
                  "is2FAEnabled": { "type": "boolean", "example": true },
                  "currencyPreference": { "type": "string", "example": "USD" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Profile updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "message": { "type": "string", "example": "Profile updated successfully" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/wallet/summary": {
      "get": {
        "summary": "Fetch total portfolio, balances, earnings, and 24h percentage",
        "tags": ["3. Dashboard, Wallet & Market Metrics"],
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Wallet summary overview",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": {
                      "type": "object",
                      "properties": {
                        "totalPortfolio": { "type": "number", "example": 48250.00 },
                        "availableBalance": { "type": "number", "example": 12450.00 },
                        "activeInvestments": { "type": "number", "example": 35800.00 },
                        "totalEarnings": { "type": "number", "example": 8420.50 },
                        "growth24h": { "type": "number", "example": 4.82 },
                        "currency": { "type": "string", "example": "USD" }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analytics/chart": {
      "get": {
        "summary": "Historical portfolio performance data points for charts",
        "tags": ["3. Dashboard, Wallet & Market Metrics"],
        "parameters": [
          {
            "name": "period",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["1D", "1W", "1M", "1Y", "ALL"],
              "default": "1M"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Chart data series",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "timestamp": { "type": "string", "example": "2026-08-30T00:00:00Z" },
                          "value": { "type": "number", "example": 48250.00 }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/investments": {
      "get": {
        "summary": "Get user's active, in-progress, and matured investments",
        "tags": ["5. Investments & Yield Matrix"],
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": {
            "description": "User investments list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "investments": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": { "type": "string", "example": "inv_7721" },
                          "planName": { "type": "string", "example": "Apex Starter Tier" },
                          "amount": { "type": "number", "example": 5000.00 },
                          "roi": { "type": "string", "example": "15%" },
                          "progress": { "type": "number", "example": 74.2 },
                          "projectedReturn": { "type": "number", "example": 5750.00 },
                          "status": { "type": "string", "example": "active" },
                          "startDate": { "type": "string", "example": "2026-08-25T00:00:00Z" },
                          "maturityDate": { "type": "string", "example": "2026-09-01T00:00:00Z" }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create new investment allocated from available wallet balance",
        "tags": ["5. Investments & Yield Matrix"],
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["planId", "planName", "amount", "roi"],
                "properties": {
                  "planId": { "type": "string", "example": "starter" },
                  "planName": { "type": "string", "example": "Apex Starter Tier" },
                  "amount": { "type": "number", "example": 5000.00 },
                  "roi": { "type": "string", "example": "15%" }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Investment position created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "investment": {
                      "type": "object",
                      "properties": {
                        "id": { "type": "string", "example": "inv_7722" },
                        "planName": { "type": "string", "example": "Apex Starter Tier" },
                        "amount": { "type": "number", "example": 5000.00 },
                        "roi": { "type": "string", "example": "15%" },
                        "status": { "type": "string", "example": "active" },
                        "startDate": { "type": "string", "example": "2026-08-30T21:35:00Z" }
                      }
                    },
                    "newAvailableBalance": { "type": "number", "example": 7450.00 }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/investments/{id}/settle": {
      "post": {
        "summary": "Withdraw/settle a 100% completed investment into main wallet balance",
        "tags": ["5. Investments & Yield Matrix"],
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "example": "inv_7721" }
          }
        ],
        "responses": {
          "200": {
            "description": "Settlement credited to balance",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "settlement": {
                      "type": "object",
                      "properties": {
                        "investmentId": { "type": "string", "example": "inv_7721" },
                        "payoutAmount": { "type": "number", "example": 5750.00 },
                        "creditedBalance": { "type": "number", "example": 54000.00 },
                        "transactionId": { "type": "string", "example": "tx_settle_99402" },
                        "status": { "type": "string", "example": "completed" }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/notifications": {
      "get": {
        "summary": "Fetch list of notifications and announcements for the user",
        "tags": ["8. Notifications & Alerts"],
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": {
            "description": "User notifications",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "notifications": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": { "type": "string", "example": "notif_001" },
                          "title": { "type": "string", "example": "Yield Cycle Complete" },
                          "message": { "type": "string", "example": "Your Apex Starter Tier position has matured with +15% ROI." },
                          "type": { "type": "string", "example": "yield" },
                          "isRead": { "type": "boolean", "example": false },
                          "createdAt": { "type": "string", "example": "2026-08-30T20:00:00Z" }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/notifications/{id}/read": {
      "patch": {
        "summary": "Mark a specific notification as read",
        "tags": ["8. Notifications & Alerts"],
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "example": "notif_001" }
          }
        ],
        "responses": {
          "200": {
            "description": "Marked read status",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "message": { "type": "string", "example": "Notification marked as read" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/notifications/mark-all-read": {
      "post": {
        "summary": "Mark all user notifications as read",
        "tags": ["8. Notifications & Alerts"],
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Marked all notifications as read",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "message": { "type": "string", "example": "All notifications marked as read" }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}